requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aioredis==1.3.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
databases==0.4.3
fastapi==0.65.1
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
idna==2.8
multidict==5.1.0
mysql-connector-python==8.0.25
protobuf==3.17.3
pydantic==1.8.2
PyMySQL==0.9.3
python-dotenv==0.18.0
requests==2.22.0
Shapely==1.7.1
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.25.11
uvicorn==0.13.4
yarl==1.6.3
